+++ src/components/picker/DatePicker/index.ts
@@ -1,0 +1,1 @@
+export { default as DatePicker } from './DatePicker';
