+++ src/components/picker/index.ts
@@ -1,2 +1,2 @@
-export { default as DatePicker } from './DatePicker';
+export { DatePicker } from './DatePicker';
 export { default as TimePicker } from './TimePicker';
